eslint config
pre-reorganization made of the folders, and configuration made of Eslint still needs to solve a bug in the token =, but I'm already working on it
