+++ .prettierrc.json
@@ -1,7 +1,7 @@
 {
-    "trailingComma": "all",
-    "tabWidth": 4,
-    "semi": false,
-    "singleQuote": true
-  }
-  
+  "semi": true,
+  "trailingComma": "all",
+  "singleQuote": true,
+  "printWidth": 80,
+  "tabWidth": 2
+}
