+++ package.json
@@ -33,5 +33,13 @@
     "react-icons": "^3.10.0",
     "react-markdown": "^6.0.2",
     "yup": "^0.32.8"
+  },
+  "devDependencies": {
+    "eslint": "^8.23.0",
+    "eslint-config-prettier": "^8.5.0",
+    "eslint-plugin-prettier": "^4.2.1",
+    "eslint-plugin-react": "^7.31.1",
+    "eslint-plugin-react-hooks": "^4.6.0",
+    "prettier": "^2.7.1"
   }
 }
