--- pages/empresas.js
@@ -1,7 +1,0 @@
-
-const Empresas = () => {
-
-    return <><h1>Empresas</h1></>
-}
-
-export default Empresas;
