--- pages/index.js
@@ -1,6 +1,0 @@
-const Painel = () => {
-    
-    return <h1>Painel</h1>
-};
-
-export default Painel;
