--- pages/perfil.js
@@ -1,6 +1,0 @@
-const Perfil = () => {
-    
-    return <h1>Perfil</h1>
-}
-
-export default Perfil;
