--- pages/pessoas.js
@@ -1,6 +1,0 @@
-const Pessoas = () => {
-    
-    return <h1>Hello World pessoas</h1>
-}
-
-export default Pessoas;
