--- pages/produtos.js
@@ -1,8 +1,0 @@
-import Economica from "../components/box-models/economica";
-
-const Produtos = () => {
-
-    return <h1>Produtos</h1>
-};
-
-export default Produtos;
