--- pages/tarefas.js
@@ -1,6 +1,0 @@
-const Tarefas = () => {
-    
-    return <h1>Tarefas</h1>
-}
-
-export default Tarefas;
