+++ src/components/elements/head.js
@@ -1,0 +1,13 @@
+/* eslint-disable react/prop-types */
+import Head from 'next/head';
+
+const AppHead = (props) => {
+  /* Favicon */
+  return (
+    <Head>
+      <link rel="shortcut icon" href={props.favicon} />
+    </Head>
+  );
+};
+
+export default AppHead;
