+++ src/pages/_document.js
@@ -1,9 +1,7 @@
-import Document, { Html, Head, Main, NextScript } from 'next/document'
+import Document, { Html, Head, Main, NextScript } from 'next/document';
 
 export default class MyDocument extends Document {
-  
   render() {
-
     return (
       <Html lang="pt-BR">
         <Head />
@@ -12,6 +10,6 @@ export default class MyDocument extends Document {
           <NextScript />
         </body>
       </Html>
-    )
+    );
   }
 }
