+++ src/pages/empresas.js
@@ -1,0 +1,9 @@
+const Empresas = () => {
+  return (
+    <>
+      <h1>Empresas</h1>
+    </>
+  );
+};
+
+export default Empresas;
