+++ src/pages/index.js
@@ -1,0 +1,5 @@
+const Painel = () => {
+  return <h1>Painel</h1>;
+};
+
+export default Painel;
