+++ src/pages/negocios.js
@@ -1,6 +1,5 @@
-
 const Negocios = () => {
-  return <h1>Negócios</h1>
+  return <h1>Negócios</h1>;
 };
 
 export default Negocios;
