+++ src/pages/perfil.js
@@ -1,0 +1,5 @@
+const Perfil = () => {
+  return <h1>Perfil</h1>;
+};
+
+export default Perfil;
