+++ src/pages/pessoas.js
@@ -1,0 +1,5 @@
+const Pessoas = () => {
+  return <h1>Hello World pessoas</h1>;
+};
+
+export default Pessoas;
