+++ src/pages/tarefas.js
@@ -1,0 +1,5 @@
+const Tarefas = () => {
+  return <h1>Tarefas</h1>;
+};
+
+export default Tarefas;
